refactor: modifyExpression parametrized
The function is now directly used to updated the parsed prometheus expression and adding/updating either the "namespace" / "exported_namespace" labels, which are the first-class citizens of the prometheus-auth federation hijacking mechanism.

This makes the cloning function redundant; we now just use prometheus parser to simply do the same job twice; once to add the namespace and once to add the exported_namespace to the same query. This makes the code simpler to understand and to maintain, in case a new label should be needed.